Renames bitcoin related files.
