+++ tool/bitcoin/coinbaseData.js
